pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thinkmark"
version = "0.2.0"
description = "Documentation to LLM pipeline (scrape, convert, annotate)"
readme = "README.md"
authors = [
    { name = "ThinkMark Team" },
]
requires-python = ">=3.12,<4.0"
license = { text = "MIT" } 
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "scrapy>=2.11,<3.0",
    "python-slugify>=8.0,<9.0",
    "beautifulsoup4>=4.12.3,<5.0",
    "lxml>=5.2.1,<6.0",
    "html2text>=2024.2.26,<2025.0",
    "jsonlines>=4.0.0,<5.0",
    "tqdm>=4.66.2,<5.0",
    "scikit-learn>=1.6.1,<2.0",
    "numpy>=1.26.4,<2.0",
    "openrouter>=1.0,<2.0",
    "openai>=1.75.0,<2.0",
    "typer>=0.15.2,<0.16.0",
    "click>=8.1.3,<9.0",
    "rich>=13.7.0,<14.0",
    "python-dotenv>=1.0.0,<2.0",
    "PyYAML>=6.0,<7.0",
    "fastmcp>=2.3.0,<3.0",
    "nest-asyncio>=1.5.8,<2.0",
    "faiss-cpu>=1.11.0",
]

[project.optional-dependencies]
mcp = [
    # FastMCP already includes necessary dependencies for running the MCP server
    # We keep uvicorn for HTTP server capabilities
    "uvicorn>=0.23.2,<0.24.0",
]
dev = [
    "pytest>=8.2.0,<9.0",
    "ruff>=0.1.3,<0.2.0",
    "black>=24.4.2,<25.0",
]

[project.scripts]
thinkmark = "thinkmark.cli:app"
thinkmark-mcp = "thinkmark.mcp.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"
